slackline: implement ctrl+w and ctrl+l
ctrl+w is part of the canonical line disciplines on most systems, thus
it is convenient to implement it.

ctrl+l is part of the emacs-mode, but generic and useful enough to be
implemented regardless of the current mode.

Additionally, we now check for control key combinations. This makes it
possible to ignore "invalid" key combinations. It also helps with the
implementation of an emacs-mode. Previously, pressing something like ctrl+f
would insert invisible characters, which is very irritating.

Last, it isn't weird that ctrl+u emits a NAK, since ctrl key combinations
map the letters A-Z to the ASCII codes 1-26. "u" is the 21. character in
the alphabet and NAK has the ASCII code 21.